Update README.md
There wasn't an issues page, so I'm assuming that this will get rejected as a PR :) But I wanted to draw attention to contributing wholesale projects to open source and collaborations with foundations like the CNCF to bring projects more governance and support! This feels like a critical part of the story to include.